requirements.txt
Django==4.0.4
django-appconf==1.0.5
django-bootstrap-modal-forms==2.2.0
django-crispy-forms==1.14.0
pandas==1.4.2
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary
python-dotenv
Pillow==11.1.0
cloudinary
django-cloudinary-storage
urllib3==1.26.18